Project/HealthBotV2/requirements.txt
# School Management System - Consolidated Requirements
# Updated: July 12, 2025

# AI and ML - Using latest OpenAI version
openai==1.95.0
tiktoken>=0.5.0

# Database connections
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0
PyMySQL>=1.0.0

# API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Authentication and security
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4

# Environment management
python-dotenv>=1.0.0

# Vector database
pinecone

# Utilities
pydantic>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# PDF processing
PyPDF2>=3.0.0

# LangChain for agent-based query processing
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0

# HTTP client for testing
httpx>=0.24.0

# Additional utilities
typing-extensions>=4.0.0